requirements.txt
fastapi
uvicorn[standard]
joblib
pandas
scikit-learn
xgboost
python-multipart
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pydantic